+++ examples/typegoose/object-id.scalar.ts
@@ -4,16 +4,22 @@ import { ObjectId } from "mongodb";
 export const ObjectIdScalar = new GraphQLScalarType({
   name: "ObjectId",
   description: "Mongo object id scalar type",
-  parseValue(value: string) {
-    return new ObjectId(value); // value from the client input variables
-  },
-  serialize(value: ObjectId) {
-    return value.toHexString(); // value sent to the client
-  },
-  parseLiteral(ast) {
-    if (ast.kind === Kind.STRING) {
-      return new ObjectId(ast.value); // value from the client query
+  serialize(value: unknown): string {
+    if (!(value instanceof ObjectId)) {
+      throw new Error("ObjectIdScalar can only serialize ObjectId values");
     }
-    return null;
+    return value.toHexString();
+  },
+  parseValue(value: unknown): ObjectId {
+    if (typeof value !== "string") {
+      throw new Error("ObjectIdScalar can only parse string values");
+    }
+    return new ObjectId(value);
+  },
+  parseLiteral(ast): ObjectId {
+    if (ast.kind !== Kind.STRING) {
+      throw new Error("ObjectIdScalar can only parse string values");
+    }
+    return new ObjectId(ast.value);
   },
 });