+++ src/scalars/timestamp.ts
@@ -1,13 +1,10 @@
 import { Kind, GraphQLScalarType } from "graphql";
 
-function parseValue(value: string | null) {
-  if (value === null) {
-    return null;
-  }
+function convertTimestampToDate(value: number) {
   try {
     return new Date(value);
   } catch (err) {
-    return null;
+    throw new Error("Provided date numeric value is invalid and cannot be parsed");
   }
 }
 
@@ -18,18 +15,27 @@ export const GraphQLTimestamp = new GraphQLScalarType({
     "Type represents date and time as number of milliseconds from start of UNIX epoch.",
   serialize(value: Date) {
     if (!(value instanceof Date)) {
-      throw new Error(`Unable to serialize value '${value}' as it's not instance of 'Date'`);
+      throw new Error(`Unable to serialize value '${value}' as it's not an instance of 'Date'`);
     }
     return value.getTime();
   },
-  parseValue,
-  parseLiteral(ast) {
-    if (ast.kind === Kind.INT) {
-      const num = parseInt(ast.value, 10);
-      return new Date(num);
-    } else if (ast.kind === Kind.STRING) {
-      return parseValue(ast.value);
+  parseValue(value: unknown) {
+    if (typeof value !== "number") {
+      throw new Error(
+        `Unable to parse value '${value}' as GraphQLTimestamp scalar supports only number values`,
+      );
     }
-    return null;
+
+    return convertTimestampToDate(value);
+  },
+  parseLiteral(ast) {
+    if (ast.kind !== Kind.INT) {
+      throw new Error(
+        `Unable to parse literal value of kind '${ast.kind}' as GraphQLTimestamp scalar supports only '${Kind.INT}' ones`,
+      );
+    }
+
+    const num = Number.parseInt(ast.value, 10);
+    return convertTimestampToDate(num);
   },
 });